requirements.txt
et-xmlfile==1.1.0
lxml==4.8.0
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
PyQt6-Qt6==6.3.0
PyQt6-sip==13.3.1
PySide6==6.3.0
PySide6-Addons==6.3.0
PySide6-Essentials==6.3.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
shiboken6==6.3.0
# six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-103/six-1.15.0-py2.py3-none-any.whl
# export PYTHONPATH="/users/linusbolls/projects/docx-templater/lib"
# export PREFIX_PATH="/users/linusbolls/projects/docx-templater/lib"
# pip3 install --prefix=$PREFIX_PATH --ignore-installed -r requirements.txt